Use `get_logger` fixture to test exception logged in debug.
- moves get_logger fixture to conftest.py for middleware tests
- patches app logger in tests with one that propagates so caplog works.
